Use uint64

+++ mass-deployer/pkg/mass-deployer/deployer.go
@@ -37,7 +37,7 @@ func RunDeployer(cfg Config, ctx context.Context, output string) error {
 		log.Info().Msgf("running deployer for node group %s", nodeGroup.Name)
 		firstTrial := true
 
-		if err := retry.Do(ctx, retry.WithMaxRetries(uint64(cfg.MaxRetries), retry.NewConstant(1*time.Nanosecond)), func(ctx context.Context) error {
+		if err := retry.Do(ctx, retry.WithMaxRetries(cfg.MaxRetries, retry.NewConstant(1*time.Nanosecond)), func(ctx context.Context) error {
 			if !firstTrial {
 				log.Debug().Msgf("retrying to deploy node group %s", nodeGroup.Name)
 			}
